# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Game Designer API
  version: 0.1.0
  description: |
    HTTP contract for assets, sessions, projects, and auth. Session cookie (HttpOnly) is the auth mechanism; browser sends Cookie automatically.
servers:
  - url: /api
    description: Next.js rewrites /api to Go; RSC may call Go directly (API_URL). MSW in tests uses same /api paths.
...
paths:
  /me:
    get:
      tags: [me]
      summary: Current user
      security: []
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401":
          description: Not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
    patch:
      tags: [me]
      summary: Update profile (displayName only in v1)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MePatch" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/register:
    post:
      tags: [auth]
      summary: Create account; sets session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: Email or username already in use
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
  /auth/login:
    post:
      tags: [auth]
      summary: Log in; sets session cookie
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /users/{username}:
    get:
      tags: [users]
      summary: Public profile (no email)
      parameters:
        - name: username
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserPublic" }
        "404":
          $ref: "#/components/responses/NotFound"
  /assets:
    get:
      tags: [assets]
      summary: List assets (cursor pagination)
      parameters:
        - name: scope
          in: query
          required: true
          schema:
            type: string
            enum: [public, private]
          description: public = explore; private = my library
        - name: cursor
          in: query
          schema: { type: string, nullable: true }
        - name: limit
          in: query
          schema: { type: integer, default: 24, minimum: 1, maximum: 48 }
        - name: sort
          in: query
          schema:
            type: string
            enum: [createdAt_desc, createdAt_asc, updatedAt_desc]
            default: createdAt_desc
        - name: img
          in: query
          schema:
            type: string
            enum: [no]
          description: Set to `no` to return only assets that have no `asset_images` rows (image-less). Omit to list all.
        - name: authorUsername
          in: query
          schema: { type: string }
          description: Only with `scope=public`. Filter to public assets whose author matches this username. Returns 400 if used with `scope=private`.
      responses:
        "200":
          description: Page of assets
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedAssets" }
    post:
      tags: [assets]
      summary: Create asset (private by default)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetCreate" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /assets/{id}:
    get:
      tags: [assets]
      summary: Get asset by id
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [assets]
      summary: Update editable fields (private assets only; public = read-only)
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AssetUpdate" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400":
          description: Public asset — content frozen after publish
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [assets]
      summary: Soft-delete a private asset (owner only); public assets rejected
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "204":
          description: Asset marked deleted (ghost); removed from library lists
        "400":
          description: Public or invalid state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /assets/{id}/publish:
    post:
      tags: [assets]
      summary: Publish private asset to public (irreversible content freeze per product rules)
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /assets/{id}/fork:
    post:
      tags: [assets]
      summary: Fork asset into current user's private library
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "404":
          description: Source missing or not forkable
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
  /assets/{id}/images:
    post:
      tags: [assets]
      summary: Request image generation for private asset only (public = read-only)
      parameters:
        - $ref: "#/components/parameters/AssetId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                extraPrompt: { type: string, nullable: true }
      responses:
        "202":
          description: Job accepted; new AssetImage pending
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AssetImage" }
        "400":
          description: Public asset — no new images after publish
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /assets/{id}/images/{imageId}:
    delete:
      tags: [assets]
      summary: Remove one image from a private asset (public = read-only; cover_id cleared if needed)
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - name: imageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Image removed
        "400":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorBody" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /assets/{id}/forks:
    get:
      tags: [forks]
      summary: Lazy fork neighbourhood (v1 depth=1 only)
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - name: direction
          in: query
          required: true
          schema:
            type: string
            enum: [upstream, downstream]
        - name: depth
          in: query
          schema: { type: integer, default: 1, minimum: 1, maximum: 1 }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForkPage" }
        "404":
          $ref: "#/components/responses/NotFound"
  /assets/{id}/fork-graph:
    get:
      tags: [forks]
      summary: Bounded fork subgraph for graph view (focus asset must be readable like GET /assets/{id})
      parameters:
        - $ref: "#/components/parameters/AssetId"
        - name: maxUpstream
          in: query
          schema: { type: integer, default: 64, minimum: 1, maximum: 128 }
        - name: downstreamDepth
          in: query
          schema: { type: integer, default: 2, minimum: 0, maximum: 12 }
        - name: maxNodes
          in: query
          schema: { type: integer, default: 200, minimum: 1, maximum: 500 }
        - name: expandFrom
          in: query
          description: If set, return direct children of this asset only (lazy expand); focus id in path is still the page focus for cache keys
          schema: { type: string, format: uuid }
        - name: childLimit
          in: query
          description: Max children returned per expandFrom request
          schema: { type: integer, default: 48, minimum: 1, maximum: 96 }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForkGraph" }
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions:
    get:
      tags: [sessions]
      summary: List AI material sessions
      description: Returns only sessions not bound to a game project (`project_id` null). Project design threads use `GET /projects/{projectId}/sessions`.
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SessionSummary" }
    post:
      tags: [sessions]
      summary: Create session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                stagingGroupId: { type: string, description: optional session staging group to attach }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
  /session-staging-groups:
    get:
      tags: [session-staging-groups]
      summary: List session staging groups (independent vs shared draft pool)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SessionStagingGroup" }
    post:
      tags: [session-staging-groups]
      summary: Create session staging group
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                draftStaging: { type: string, enum: [independent, shared], default: independent }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionStagingGroup" }
  /session-staging-groups/{groupId}/drafts:
    parameters:
      - name: groupId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [session-staging-groups]
      summary: List all staging draft rows for this group (shared pool or union of per-session rows)
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DraftAsset" }
        "404":
          $ref: "#/components/responses/NotFound"
  /session-staging-groups/{groupId}:
    parameters:
      - name: groupId
        in: path
        required: true
        schema: { type: string }
    patch:
      tags: [session-staging-groups]
      summary: Update name or draft staging mode (mode change requires empty staging in conflict paths)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                draftStaging: { type: string, enum: [independent, shared] }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionStagingGroup" }
    delete:
      tags: [session-staging-groups]
      summary: Delete group (sessions unlinked; group-scoped drafts removed)
      responses:
        "204": { description: No content }
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{sessionId}:
    get:
      tags: [sessions]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [sessions]
      summary: Update title and/or assign session to a staging group
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                stagingGroupId: { type: string, nullable: true, description: set null to remove from group }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [sessions]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: Archived / deleted }
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{sessionId}/chat:
    post:
      tags: [stream]
      summary: Start assistant reply stream (SSE, JSONL body lines)
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
      responses:
        "200":
          description: text/event-stream of JSONL lines
          content:
            text/event-stream:
              schema: { type: string, format: binary }
  /sessions/{sessionId}/drafts:
    post:
      tags: [sessions]
      summary: Add a user-authored draft (name + description) to session staging; not auto-created by chat stream
      parameters:
        - $ref: "#/components/parameters/SessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DraftAsset" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{sessionId}/drafts/{tempId}:
    patch:
      tags: [sessions]
      summary: Update a staged draft (name + description)
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/DraftTempId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name: { type: string }
                description: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DraftAsset" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [sessions]
      summary: Remove a draft from session staging
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/DraftTempId"
      responses:
        "204": { description: Removed }
        "404":
          $ref: "#/components/responses/NotFound"
  /sessions/{sessionId}/drafts/{tempId}/export-to-library:
    post:
      tags: [sessions]
      summary: Export one staged draft to private library and remove it from staging (atomic)
      description: |
        Uses draft name/description from the staging row. For shared group staging, `sessionId` must be any chat session in that group (same as PATCH/DELETE draft).
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/DraftTempId"
      responses:
        "201":
          description: Created private asset; draft row deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Asset" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects:
    get:
      tags: [projects]
      summary: List projects
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectSummary" }
    post:
      tags: [projects]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
  /projects/{projectId}:
    get:
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectPatch" }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/sessions:
    get:
      tags: [projects]
      summary: List design chat sessions for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProjectSessionSummary" }
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [projects]
      summary: Create a design chat session under a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
      responses:
        "201":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/assets:
    post:
      tags: [projects]
      summary: Link a private library asset to a project (AI context)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [assetId]
              properties:
                assetId: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/{projectId}/assets/{assetId}:
    delete:
      tags: [projects]
      summary: Remove linked asset from project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: assetId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectDetail" }
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    AssetId:
      name: id
      in: path
      required: true
      schema: { type: string }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { type: string }
    DraftTempId:
      name: tempId
      in: path
      required: true
      schema: { type: string }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
    NotFound:
      description: Not found (or 404 for private resource per product rules)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        error: { type: string, description: Human-readable message for UI }
        code: { type: string, description: Stable machine code }
      required: [error]
    Me:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string, nullable: true }
      required: [id, username]
    MePatch:
      type: object
      properties:
        displayName: { type: string, nullable: true }
    RegisterRequest:
      type: object
      required: [email, password, username]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        username: { type: string, minLength: 1, maxLength: 32 }
        displayName: { type: string, nullable: true }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    UserPublic:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string, nullable: true }
      required: [id, username]
...
    AssetImage:
      type: object
      properties:
        id: { type: string }
        assetId: { type: string }
        url: { type: string, description: "CDN template, append ?w=" }
        extraPrompt: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        generationStatus:
          type: string
          enum: [pending, done, failed]
      required: [id, assetId, url, createdAt, generationStatus]
    Asset:
      oneOf:
        - $ref: "#/components/schemas/AssetFull"
        - $ref: "#/components/schemas/AssetDeletedGhost"
      discriminator:
        propertyName: visibility
        mapping:
          deleted: "#/components/schemas/AssetDeletedGhost"
          private: "#/components/schemas/AssetFull"
          public: "#/components/schemas/AssetFull"
...
    AssetCreate:
      type: object
      required: [name, description]
      properties:
        name: { type: string }
        description: { type: string }
        annotation: { type: string, nullable: true }
        sessionId: { type: string, description: optional link to originating AI session }
        forkedFromId: { type: string, nullable: true }
    AssetUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        annotation: { type: string, nullable: true }
        coverImageId: { type: string, nullable: true }
        # publish/delete via dedicated routes
    PaginatedAssets:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Asset" }
        nextCursor: { type: string, nullable: true }
        total: { type: number, nullable: true }
      required: [items, nextCursor]
...
    ForkPage:
      type: object
      properties:
        direction: { type: string, enum: [upstream, downstream] }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/ForkNode" }
      required: [direction, nodes]
...
    ForkGraph:
      type: object
      properties:
        focusAssetId: { type: string }
        nodes:
          type: array
          items: { $ref: "#/components/schemas/ForkGraphNode" }
        truncated: { type: boolean, description: true if maxNodes or childLimit cut off results }
      required: [focusAssetId, nodes, truncated]
...
    SessionStagingGroup:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        position: { type: integer }
        draftStaging: { type: string, enum: [independent, shared] }
        createdAt: { type: string, format: date-time }
      required: [id, name, position, draftStaging, createdAt]
    SessionSummary:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        updatedAt: { type: string, format: date-time }
        draftAssetCount: { type: integer }
        stagingGroup: { $ref: "#/components/schemas/SessionStagingGroupRef", nullable: true }
      required: [id, title, updatedAt, draftAssetCount]
    SessionDetail:
      allOf:
        - { $ref: "#/components/schemas/SessionSummary" }
        - type: object
          properties:
            projectId:
              type: string
              nullable: true
              description: Present when this thread is a project design session (hidden from global `/sessions` list)
            messages:
              type: array
              items: { $ref: "#/components/schemas/ChatMessage" }
            draftAssets:
              type: array
              items: { $ref: "#/components/schemas/DraftAsset" }
...
    DraftAsset:
      type: object
      properties:
        tempId: { type: string }
        name: { type: string }
        description: { type: string }
        done: { type: boolean, default: false }
        ownerSessionId:
          type: string
          description: Present when draftStaging is independent — session that owns this row (for PATCH/DELETE paths)
        ownerSessionTitle: { type: string }
      required: [tempId, name, description, done]
    ProjectSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        updatedAt: { type: string, format: date-time }
      required: [id, name, updatedAt]
...
    ProjectDetail:
      allOf:
        - { $ref: "#/components/schemas/ProjectSummary" }
        - type: object
          properties:
            canvasDocument:
              type: object
              description: tldraw document JSON; only asset refs are semantic
              additionalProperties: true
            linkedAssets:
              type: array
              items: { $ref: "#/components/schemas/ProjectLinkedAsset" }
            designDocument:
              type: string
              description: Markdown game design document (GDD)
          required: [linkedAssets, designDocument]
...